Corrected logic for extracting LED state data
In short, the LED states are held in one byte configured for USB HID
protocol, and need to be reconfigured to be understood by IBM's
protocols (for sending to keyboards that have LEDs) and by QMK (for
sending to the converter to account for keyboards that have no LEDs).

I may further optimise this slightly, but it appears to be working well
already, and I have confirmed no ill effects of the Unicomp Pause/Break
workaround on my New Model M as well.